Network diagnostics:  - move logic into cluster  - Update event logging to not depend on logevent and use the context.
