+++ src/app/clusters/wifi-network-diagnostics-server/tests/TestWiFiNetworkDiagnosticsCluster.cpp
@@ -51,7 +51,8 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, CompileTest)
     };
 
     NullProvider nullProvider;
-    WiFiDiagnosticsServerCluster cluster(kRootEndpointId, nullProvider, {}, BitFlags<WiFiNetworkDiagnostics::Feature>(0));
+    WiFiDiagnosticsServerCluster cluster(kRootEndpointId, nullProvider, WiFiDiagnosticsServerCluster::OptionalAttributeSet(),
+                                         BitFlags<WiFiNetworkDiagnostics::Feature>(0));
 
     // Essentially say "code executes"
     ASSERT_EQ(cluster.GetClusterFlags({ kRootEndpointId, WiFiNetworkDiagnostics::Id }), BitFlags<ClusterQualityFlags>());
@@ -66,7 +67,8 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, AttributesTest)
         };
 
         NullProvider nullProvider;
-        WiFiDiagnosticsServerCluster cluster(kRootEndpointId, nullProvider, {}, BitFlags<WiFiNetworkDiagnostics::Feature>(0));
+        WiFiDiagnosticsServerCluster cluster(kRootEndpointId, nullProvider, WiFiDiagnosticsServerCluster::OptionalAttributeSet(),
+                                         BitFlags<WiFiNetworkDiagnostics::Feature>(0));
 
         // without any enabled attributes, no commands are accepted
         ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> commandsBuilder;
@@ -109,7 +111,7 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, AttributesTest)
 
         ErrorCountsProvider errorCountsProvider;
         WiFiDiagnosticsServerCluster cluster(
-            kRootEndpointId, errorCountsProvider, {},
+            kRootEndpointId, errorCountsProvider, WiFiDiagnosticsServerCluster::OptionalAttributeSet(),
             BitFlags<WiFiNetworkDiagnostics::Feature>(WiFiNetworkDiagnostics::Feature::kErrorCounts));
 
         ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> commandsBuilder;
@@ -217,7 +219,7 @@ TEST_F(TestWiFiNetworkDiagnosticsCluster, AttributesTest)
         features.Set(WiFiNetworkDiagnostics::Feature::kErrorCounts);
         features.Set(WiFiNetworkDiagnostics::Feature::kPacketCounts);
         WiFiDiagnosticsServerCluster cluster(
-            kRootEndpointId, allProvider, WiFiDiagnosticsServerLogic::OptionalAttributeSet().Set<CurrentMaxRate::Id>(), features);
+            kRootEndpointId, allProvider, WiFiDiagnosticsServerCluster::OptionalAttributeSet().Set<CurrentMaxRate::Id>(), features);
 
         ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> commandsBuilder;
         ASSERT_EQ(cluster.AcceptedCommands(ConcreteClusterPath(kRootEndpointId, WiFiNetworkDiagnostics::Id), commandsBuilder),
